refactor(filestorage): rename ports.py to contracts.py (#5365)
