+++ infrastructure/filestorage/engine.py
@@ -32,10 +32,10 @@ from enum import StrEnum
 from pathlib import Path
 
 from config.constants.filestorage import DEFAULT_MAX_PARALLEL_UPLOADS
+from infrastructure.filestorage.contracts import ObjectStore, RemoteObject
 from infrastructure.filestorage.enums import SyncDirection
 from infrastructure.filestorage.errors import RemoteSyncConfigError, UnsyncablePathError
 from infrastructure.filestorage.exclusions import NO_EXCLUSIONS, ExclusionRules
-from infrastructure.filestorage.ports import ObjectStore, RemoteObject
 from infrastructure.filestorage.syncable import SyncRoot, resolved_roots, syncable_roots
 
 logger = logging.getLogger(__name__)
