+++ tests/filestorage/test_multi_provider_support.py
@@ -14,10 +14,10 @@ from pathlib import Path
 import pytest
 
 from infrastructure.filestorage.config import RemoteSyncConfig
+from infrastructure.filestorage.contracts import RemoteObject
 from infrastructure.filestorage.engine import content_tag, run_sync
 from infrastructure.filestorage.enums import SyncRootName
 from infrastructure.filestorage.errors import RemoteSyncConfigError
-from infrastructure.filestorage.ports import RemoteObject
 from infrastructure.filestorage.providers.registry import (
     build_object_store,
     register_object_store,
@@ -74,8 +74,8 @@ def roots(tmp_path: Path) -> tuple[SyncRoot, ...]:
     memory = tmp_path / "memory"
     sessions.mkdir()
     memory.mkdir()
-    (sessions / "a.jsonl").write_text('{"turn": 1}\n', encoding="utf-8")
-    (memory / "fact.md").write_text("remembered\n", encoding="utf-8")
+    (sessions / "a.jsonl").write_bytes(b'{"turn": 1}\n')
+    (memory / "fact.md").write_bytes(b"remembered\n")
     return (
         SyncRoot(name=SyncRootName.SESSIONS, path=sessions),
         SyncRoot(name=SyncRootName.MEMORY, path=memory),
